+++ external-services/db-postgres-repository/src/main/kotlin/com/br/gook/repositories/impl/LocalRepositoryImpl.kt
@@ -23,12 +23,12 @@ class LocalRepositoryImpl(
         return localRepository.save(localPort.toEntity()).toPort()
     }
 
-    override fun findLocalByIdOrThrow(localId: Int): LocalOutputPort {
+    override fun findLocalByIdOrThrow(localId: Long): LocalOutputPort {
         return localRepository.findById(localId).takeIf { it.isPresent }?.get()?.toPort()
             ?: throw Exception("LocalRepositoryImpl.findLocalByIdOrThrow - Error to find Local - localId: $localId")
     }
 
-    override fun deleteLocal(localId: Int) {
+    override fun deleteLocal(localId: Long) {
         val localPort = findLocalByIdOrThrow(localId)
         localRepository.delete(localPort.toEntity())
     }
