+++ external-services/db-postgres-repository/src/main/kotlin/com/br/gook/repositories/interfaces/AddressRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface AddressRepository : JpaRepository<AddressEntity, Int> {}
+interface AddressRepository : JpaRepository<AddressEntity, Long> {}
