+++ external-services/db-postgres-repository/src/main/kotlin/com/br/gook/repositories/interfaces/CourtRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CourtRepository : JpaRepository<CourtEntity, Int> {}
+interface CourtRepository : JpaRepository<CourtEntity, Long> {}
