+++ external-services/db-postgres-repository/src/main/kotlin/com/br/gook/repositories/interfaces/SchedulerRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SchedulerRepository : JpaRepository<SchedulerEntity, Int> {}
+interface SchedulerRepository : JpaRepository<SchedulerEntity, Long> {}
