+++ use-cases/use-case-port/src/main/kotlin/com/br/gook/port/output/LocalRepositoryOutput.kt
@@ -6,7 +6,7 @@ interface LocalRepositoryOutput {
 
     fun saveLocal(localPort: LocalOutputPort): LocalOutputPort
 
-    fun findLocalByIdOrThrow(localId: Int): LocalOutputPort
+    fun findLocalByIdOrThrow(localId: Long): LocalOutputPort
 
-    fun deleteLocal(localId: Int)
+    fun deleteLocal(localId: Long)
 }
